--- user-service/src/main/java/com/winner/client/userservice/user/application/command/RefreshTokenCommand.java
@@ -1,7 +1,0 @@
-package com.winner.client.userservice.user.application.command;
-
-public record RefreshTokenCommand(
-    String refreshToken
-) {
-
-}
