+++ user-service/src/main/java/com/winner/client/userservice/user/application/dto/command/LoginCommand.java
@@ -1,4 +1,4 @@
-package com.winner.client.userservice.user.application.command;
+package com.winner.client.userservice.user.application.dto.command;
 
 public record LoginCommand(
     String userName,
