+++ user-service/src/main/java/com/winner/client/userservice/user/application/dto/command/RefreshTokenCommand.java
@@ -1,0 +1,7 @@
+package com.winner.client.userservice.user.application.dto.command;
+
+public record RefreshTokenCommand(
+    String refreshToken
+) {
+
+}
