+++ user-service/src/main/java/com/winner/client/userservice/user/presentation/controller/InternalController.java
@@ -1,9 +1,9 @@
-package com.winner.client.userservice.user.presentation;
+package com.winner.client.userservice.user.presentation.controller;
 
 import com.winner.client.global.response.ApiResponse;
 import com.winner.client.global.response.CommonSuccessCode;
 import com.winner.client.userservice.user.application.service.UserQueryService;
-import com.winner.client.userservice.user.presentation.response.UserDetailResponse;
+import com.winner.client.userservice.user.presentation.dto.response.UserDetailResponse;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
